requirements.txt
appnope==0.1.4 
asttokens==2.4.1 
async-timeout==5.0.1 
backoff==2.2.1 
beautifulsoup4==4.12.3 
bs4==0.0.2 
certifi==2024.8.30 
cffi==1.17.1 
charset-normalizer==3.4.0 
colorama==0.4.6 
comm==0.2.2 
debugpy==1.8.8 
decorator==5.1.1 
exceptiongroup==1.2.2 
executing==2.1.0 
falkordb==1.0.9 
fix-busted-json==0.0.18 
idna==3.10 
importlib-metadata==8.5.0 
ipykernel==6.29.5 
ipython==8.18.1 
jedi==0.19.2 
jupyter-client==8.6.3 
jupyter-core==5.7.2 
matplotlib-inline==0.1.7 
nest-asyncio==1.6.0 
packaging==24.2 
parso==0.8.4 
pexpect==4.9.0 
platformdirs==4.3.6 
prompt-toolkit==3.0.48 
psutil==6.1.0 
ptyprocess==0.7.0 
pure-eval==0.2.3 
pycparser==2.22 
pygments==2.18.0 
pypdf==5.1.0 
python-abc==0.2.0 
python-dateutil==2.9.0
python-docx==1.1.0 
python-dotenv==1.0.1 
pyzmq==26.2.0 
ratelimit==2.2.1 
redis==5.2.0 
requests==2.32.3 
six==1.16.0 
soupsieve==2.6 
stack-data==0.6.3 
tornado==6.4.1 
traitlets==5.14.3 
typing-extensions==4.12.2 
urllib3==2.2.3 
wcwidth==0.2.13 
zipp==3.21.0 
annotated-types==0.7.0 
anthropic==0.48.0 
anyio==4.2.0 
click==8.1.7 
cohere==5.3.1 
cryptography==42.0.2 
dataclasses-json==0.6.4 
dict-recursive-update==1.0.1 
distro==1.9.0 
# fastavro==1.9.4 
filelock==3.13.1 
h11==0.14.0 
h2==4.1.0 
hpack==4.0.0 
http-message-signatures==0.5.0 
httpcore==1.0.2 
httpx==0.27.0 
huggingface-hub==0.20.3 
hyperframe==6.0.1 
jq==1.6.0 
jsonpatch==1.33 
jsonpointer==2.4 
litellm==1.74.1 
marshmallow==3.20.2 
mistune==3.0.2 
mypy-extensions==1.0.0 
nltk==3.8.1 
numpy==1.26.4
openai==1.68.2 
orjson==3.9.14 
pdfminer-six==20221105 
pdfplumber==0.10.3 
pillow==10.2.0 
pypdfium2==4.24.0 
pydantic-core==2.18.4
pydantic==2.7.4
rapidfuzz==3.6.1 
regex==2023.12.25 
sentence-transformers==2.7.0 
sniffio==1.3.0 
tenacity==8.3.0 
tqdm==4.66.1 
typing-inspect==0.9.0 
yarl==1.9.4 
#New dependencies from Data Extractor Logic
spacy==3.7.2
PyMuPDF==1.23.8
pytesseract==0.3.10
pdf2image==1.17.0
pathlib==1.0.1
docx2python==2.0.3
lxml==4.9.3
langdetect==1.0.9
google-generativeai == 0.8.5 # For Google Gemini API
inflect==7.0.0
networkx==3.3
python-louvain==0.16
scipy==1.13.0
Flask==3.0.3
Flask-Cors==4.0.1

#Paddel OCR and Langfuse dependencies
#paddleocr==3.0.0    # use this version, stable and compatible with paddlepaddle. Should automatically install paddlex
#paddlepaddle==3.0.0 # Ensure paddlepaddle is compatible with paddleocr
tokenizers==0.15.2  
langfuse==3.2.1
groq == 0.30.0 # For Groq API
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.4
agno==1.3.0
faiss-cpu==1.12.0
leidenalg==0.10.1
python-igraph==0.10.0
